prerequisite/openCV/requirements.txt
opencv-python
Pillow
numpy
mediapipe